+++ employee-service/src/main/java/com/gabrielluciano/employeeservice/domain/dto/EmployeeResponse.java
@@ -3,11 +3,13 @@ package com.gabrielluciano.employeeservice.domain.dto;
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.gabrielluciano.employeeservice.domain.model.Employee;
 
-public record EmployeeResponse(String name, String cpf, Long baseSalary,
+import java.math.BigDecimal;
+
+public record EmployeeResponse(String name, String cpf, BigDecimal baseSalary,
                                @JsonAlias("position") PositionResponse positionResponse) {
 
     public static EmployeeResponse fromModel(Employee employee) {
-        return new EmployeeResponse(employee.getName(), employee.getCpf(), employee.getBaseSalary().longValue(),
+        return new EmployeeResponse(employee.getName(), employee.getCpf(), employee.getBaseSalary(),
                 PositionResponse.fromModel(employee.getPosition()));
     }
 }
